+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Volumes now have a “Transform Subpath” setting.
 - Added the `|address` Twig filter.
 - Added the `craft.addresses()` address query factory function for Twig.
+- Added the `users/delete-address` action.
 - Added the `users/save-address` action.
 - Added `craft\base\Element::notesFieldHtml()`.
 - Added `craft\base\Element::statusFieldHtml()`.